requirements.txt
anthropic>=0.18.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.0
aiohttp>=3.9.0
lxml>=4.9.0